// types/courseSection/enums.ts
export enum CourseSectionMessage {
  SUCCESS_CREATE = "Tạo chương học thành công",
  FAIL_CREATE = "Tạo chương học thất bại",
  
  SUCCESS_UPDATE = "Cập nhật chương học thành công",
  FAIL_UPDATE = "Cập nhật chương học thất bại",
  
  SUCCESS_DELETE = "Xóa chương học thành công",
  FAIL_DELETE = "Xóa chương học thất bại",
  
  SUCCESS_GET = "Lấy thông tin chương học thành công",
  FAIL_GET = "Lấy thông tin chương học thất bại",
  
  SECTION_NOT_FOUND = "Không tìm thấy chương học",
  SUCCESS_REORDER = "Sắp xếp lại chương học thành công",
  FAIL_REORDER = "Sắp xếp lại chương học thất bại",
  UNAUTHORIZED_ACCESS = "Bạn không có quyền truy cập chương học này"
}